+++ src/helpers/setting.ts
@@ -1,3 +1,4 @@
+import { forEach } from "chootils/dist/loops";
 import { repondMeta as meta } from "../meta";
 import { _updateRepond } from "../updating";
 import { runWhenAddingAndRemovingItems, runWhenUpdatingRepond } from "./runWhens";
@@ -16,15 +17,17 @@ export function _setState(newState: any, callback?: any) {
   }, callback);
 }
 
-export function _removeItem({ type: itemType, id: itemId }: { type: string; id: string }, callback?: any) {
+export function _removeItem({ type, id }: { type: string; id: string }, callback?: any) {
+  if (!meta.willRemoveItemsInfo[type]) meta.willRemoveItemsInfo[type] = {};
+  meta.willRemoveItemsInfo[type][id] = true;
   runWhenAddingAndRemovingItems(() => {
     // removing itemId
-    delete meta.nowState[itemType][itemId];
-    meta.itemIdsByItemType[itemType] = Object.keys(meta.nowState[itemType]);
+    delete meta.nowState[type][id];
+    meta.itemIdsByItemType[type] = Object.keys(meta.nowState[type]);
     // delete meta.currentRefs[itemType][itemId]; // now done at the end of update repond
-    meta.recordedStepEndEffectChanges.itemTypesBool[itemType] = true;
+    meta.recordedStepEndEffectChanges.itemTypesBool[type] = true;
     meta.recordedStepEndEffectChanges.somethingChanged = true;
-    meta.recordedEffectChanges.itemTypesBool[itemType] = true;
+    meta.recordedEffectChanges.itemTypesBool[type] = true;
     meta.recordedEffectChanges.somethingChanged = true;
   }, callback);
 }
@@ -33,6 +36,8 @@ export function _addItem(
   { type, id, state, refs }: { type: string; id: string; state?: any; refs?: any },
   callback?: any
 ) {
+  if (!meta.willAddItemsInfo[type]) meta.willAddItemsInfo[type] = {};
+  meta.willAddItemsInfo[type][id] = true;
   runWhenAddingAndRemovingItems(() => {
     meta.nowState[type][id] = {
       ...meta.defaultStateByItemType[type](id),
@@ -45,7 +50,7 @@ export function _addItem(
     meta.itemIdsByItemType[type].push(id);
     meta.recordedStepEndEffectChanges.itemTypesBool[type] = true;
 
-    // TODO Figure out if adding an item should record the properties as chnaged or not?
+    // TODO Figure out if adding an item should record the properties as changed or not?
 
     meta.recordedStepEndEffectChanges.itemPropsBool[type][id] = {};
     meta.recordedEffectChanges.itemPropsBool[type][id] = {};
@@ -59,5 +64,16 @@ export function _addItem(
 
     meta.recordedEffectChanges.itemIdsBool[type][id] = true;
     meta.recordedEffectChanges.somethingChanged = true;
+
+    // NOTE new items with props different to the defaults props are recorded as changed
+    const itemPropNames = meta.propNamesByItemType[type];
+    forEach(itemPropNames, (propName) => {
+      const propChangedFromDefault =
+        meta.nowState[type][id][propName] !== meta.defaultStateByItemType[type](id)[propName];
+      if (propChangedFromDefault) {
+        meta.recordedStepEndEffectChanges.itemPropsBool[type][id][propName] = true;
+        meta.recordedEffectChanges.itemPropsBool[type][id][propName] = true;
+      }
+    });
   }, callback);
 }